requirements.txt
discord-py-interactions==4.3.4
python-dotenv==0.21.0